Crate bmw_conf

The BMW Configuration crate

The Configuration crate is used by other crates in the BMW repo to build and check configurations. Configurations should generally be built using the crate::config! macro and the crate::Config::check_config function should be used to confirm the resulting configuration has only allowed values, has all required values, and has no duplicates.

Examples

 use bmw_err::*;
 use bmw_conf::*;

 fn main() -> Result<(), Error> {
     // create a simple config
     let config = config!(
         AutoRotate(true),
         MaxAgeMillis(60 * 60 * 1_000),
         FileHeader("myheader".to_string())
     );

     let res = config.check_config(
         vec![
             ConfigOptionName::AutoRotate,
             ConfigOptionName::MaxAgeMillis,
             ConfigOptionName::FileHeader
         ],
         vec![ConfigOptionName::AutoRotate]
     );

     // this configuration is ok because all fields specified are allowed (AutoRotate,
     // FileHeader, and MaxAgeMillis) and all required fields (AutoRotate) are specified.
     assert!(res.is_ok());

     // create an invalid config
     let config = config!(MaxAgeMillis(60 * 60 * 1_000), FileHeader("myheader".to_string()));

     let res = config.check_config(
         vec![
             ConfigOptionName::AutoRotate,
             ConfigOptionName::MaxAgeMillis,
             ConfigOptionName::FileHeader],
         vec![ConfigOptionName::AutoRotate]
     );

     // this configuration is invalid because AutoRotate is not specified.
     assert!(res.is_err());

     Ok(())
 }

Macros

config

The config macro allows for a configuration to be specified and checked conveniently. This macro is used throughout BMW.

Structs

ConfigBuilder

A builder struct which can be used to build configs. This is typically done using the crate::config! macro which calls this builder.

Enums

ConfigOption

Configuration options used throughout BMW via macro.

ConfigOptionName

Names of configuration options used throughout BMW via macro. This correspondes to the values in crate::ConfigOption.

Traits

Config

The config trait allows for easy construction of configurations. Configurations can be retrieved with the crate::Config::get function and configurations can be checked with the crate::Config::check_config function.