Macro bmw_conf::config

macro_rules! config {
    ( $( $config:expr ),* ) => { ... };
}

The config macro allows for a configuration to be specified and checked conveniently. This macro is used throughout BMW.

Examples

 use bmw_conf::*;
 use bmw_err::*;

 // create a config using the macro and check it
 fn main() -> Result<(), Error> {
     // create a simple config
     let config = config!(FileHeader("test".to_string()), DeleteRotation(false));

     // check it
     let res = config.check_config(
         vec![
             ConfigOptionName::FileHeader,
             ConfigOptionName::DeleteRotation
         ],
         vec![ConfigOptionName::DeleteRotation]
     );

     // this configuration is ok because both FileHeader and DeleteRotation are allowed
     // and the only required configuration 'DeleteRotation' is specified
     assert!(res.is_ok());
     Ok(())
 }